move building hash table to hash join build
