+++ src/processor/include/physical_plan/operator/hash_join/hash_join_probe.h
@@ -3,7 +3,6 @@
 #include "src/common/include/types.h"
 #include "src/common/include/vector/operations/vector_node_id_operations.h"
 #include "src/processor/include/physical_plan/operator/hash_join/hash_join_build.h"
-#include "src/processor/include/physical_plan/operator/hash_join/hash_table.h"
 #include "src/processor/include/physical_plan/operator/physical_operator.h"
 #include "src/processor/include/physical_plan/operator/tuple/data_chunks.h"
